langgraph-react/pyproject.toml
[project]
name = "langgraph-react"
version = "0.1.0"
description = ""
authors = [
    {name = "AdityaC4",email = "adityapchaudhari@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.24,<0.4.0)",
    "langgraph (>=0.3.34,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-tavily (>=0.1.5,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
